wip: dex tx

+++ lib/network/model/token_balance.dart
@@ -1,7 +1,7 @@
 class TokenBalance {
   final String hash;
   final String idToken;
-  final double balance;
+  final int balance;
   final bool isPopularToken;
 
   String get balanceDisplayRounded => (balance / 100000000).toStringAsFixed(8);
